requirements.txt
langchain
langgraph
langchain-openai
langchain-community
streamlit
fpdf
python-dotenv
typing-extensions